src/main/java/com/nexos/controller/UserController.java
package com.nexos.controller;

import java.util.List;
import com.nexos.entity.User;
import com.nexos.service.UserService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.beans.factory.annotation.Autowired;

@RestController
@RequestMapping("/user")
@CrossOrigin(origins = "*", methods = { RequestMethod.GET, RequestMethod.POST })
public class UserController {

	@Autowired
	private UserService userService;

	@GetMapping
	@RequestMapping("/listUser")
	public ResponseEntity<List<User>> list() {
		return new ResponseEntity<List<User>>(userService.findAll(), HttpStatus.OK);
	}

	@PostMapping
	@RequestMapping("/addUser")
	public ResponseEntity<String> addUser(@RequestBody User user) {
		return userService.addUser(user);
	}

}
